This is a demonstration of the State Machine Pattern (https://en.wikipedia.org/wiki/State_pattern). What that explanation doesn't say explicitly is an important point: the return type of the state execute method is the IControlState type. I defined an interface called IControlState for all states, then I define 3 types of states: one from tracking the state of the arm extension controls, one to track the arm raising/lowering state, and one to track the drive state. I'm not sure some of these are necessary, but I implemented the ArmState.
When we first implemented the arm controls, we defined 3 states: 1. arm is stationary/neither X nor Y are being pushed, 2. arm is being raised/X is being pushed, 3. arm is being lowered/Y is being pushed. That was the motive behind the xPressed and yPressed variables. It's easier to understand conceptually, but it can be a bit ambiguous, because what if X and Y are being pushed? It also creates code that can be difficult to understand because it has many if/else branches and sometimes debugging can be harder.

The alternative is the State Machine Pattern. Every state will transition to another state in reaction to certain actions performed by the user. The ArmStationary state will transition to ArmExtending state if X is pressed and to ArmRetracting if Y is pressed. ArmExtending will transition to ArmStationary when X is no longer pressed. ArmRetracting will transition to ArmStationary with Y is no longer pressed. With this arrangement, FTC2022OpMode can handle whatever is necessary by executing "armState = armState.executeState(hardwareMap, gamepad1)". Whatever object is stored in armState will do whatever needs doing with respect to whatever state the program is in, and then transition to the next state by returning an instance of the new state so the next loop will do the next loop correctly.

This can be harder to wrap your head around, but it can also be easier to debug because you can isolate what the application will do in a particular state and see what it will do without having to simulate all of the conditions in a large if/else if/.../else block. You can use what is called a "unit testing system" to instantiate the state, pass in a "mock object" that imitates a real Gamepad but you can control what it returns in each test, and verify that the state acts correctly. You can also test it in-place better by commenting out or initializing the state variables differently.